refactor: Remove unused types from block streamer
