--- block-streamer/src/rules/types/mod.rs
@@ -1,3 +1,0 @@
-pub mod events;
-pub mod indexer_rule_match;
-pub mod transactions;
